Discriminant {DataMarketFinal}	R Documentation

deter data contains at least one observation and two variables

Description

deter data contains at least one observation and two variables

Usage


Discriminant(data)

Arguments

data
data:matrix or dataframe

Value

string:OK Check,it maybe something wrong

Author(s)

Shunmao

Examples



## The function is currently defined as
function (data) 
{
    if (dim(data)[2] == 1) {
        return("Check,it maybe something wrong.")
    }
    if (dim(data)[1] == 0) {
        return("Check,it maybe something wrong.")
    }
    return("OK")
  }

[Package DataMarketFinal version 1.0 Index]