add_new_words {DataMarketFinal}	R Documentation

add self new words in jiebar

Description

contain many specific words in open data of government,so need to add to dictionary to separate more properly.

Usage


add_new_words(x)

Arguments

x
worker of jiebar package

Value

logical number:True or False which implies adding new words in worker done or not.

Author(s)

Shunmao

Examples


library(jiebaR)
seg=worker()
add_new_words(seg) #add new words in dictionary

[Package DataMarketFinal version 1.0 Index]